LPS-140124 Simplify
