+++ modules/apps/object/object-api/bnd.bnd
@@ -13,7 +13,6 @@ Export-Package:\
 	com.liferay.object.related.models,\
 	com.liferay.object.rest.context.path,\
 	com.liferay.object.scope,\
-	com.liferay.object.security.permission.resource,\
 	com.liferay.object.service,\
 	com.liferay.object.service.persistence,\
 	com.liferay.object.system,\
