--- modules/apps/object/object-api/src/main/resources/com/liferay/object/security/permission/resource/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
